refactor: 기존의 기물 도메인을 삭제하고 새로운 기물 도메인을 사용하도록 변경
