+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,41 +1,41 @@
 package chess.domain.piece;
 
 import chess.domain.board.Board;
-import chess.domain.movepattern.MovePattern;
 import chess.domain.position.Position;
 import java.util.List;
 
-public class Empty extends Piece {
+public class Empty implements Piece {
 
-    public Empty(final Type type, final Side side) {
-        super(type, side);
+    private final Type type;
+    private final Side side;
+
+    public Empty() {
+        this.type = Type.EMPTY;
+        this.side = Side.NEUTRALITY;
     }
 
     @Override
-    protected void validate(final Type type, final Side side) {
-        validateType(type);
-        validateSide(side);
-    }
-
-    private void validateType(final Type type) {
-        if (type != Type.EMPTY) {
-            throw new IllegalArgumentException("Empty의 타입이 잘못되었습니다.");
-        }
-    }
-
-    private void validateSide(final Side side) {
-        if (side != Side.NEUTRALITY) {
-            throw new IllegalArgumentException("Empty는 진영을 가질 수 없습니다.");
-        }
+    public List<Position> findMovablePosition(final Position source, final Board board) {
+        throw new UnsupportedOperationException("지원하지 않는 기능입니다.");
     }
 
     @Override
-    protected List<MovePattern> getMovePatterns() {
-        throw new UnsupportedOperationException("지원하지 않는 메서드입니다.");
+    public String name() {
+        return type.getSymbol(side);
     }
 
     @Override
-    public List<Position> findMovablePositions(final Position source, final Board board) {
-        throw new UnsupportedOperationException("지원하지 않는 메서드입니다.");
+    public Side side() {
+        return side;
+    }
+
+    @Override
+    public boolean isPawn() {
+        return false;
+    }
+
+    @Override
+    public void changePawnMoveState() {
+        throw new UnsupportedOperationException("지원하지 않는 기능입니다.");
     }
 }
