+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,39 +1,40 @@
 package chess.domain.piece;
 
+import chess.domain.board.Board;
 import chess.domain.movepattern.KnightMovePattern;
-import chess.domain.movepattern.MovePattern;
-import java.util.Arrays;
+import chess.domain.position.Position;
 import java.util.List;
 
-public class Knight extends ImmediatePiece {
+public class Knight implements Piece {
 
-    private final List<MovePattern> movePatterns;
+    private final ImmediatePiece immediatePiece;
 
-    public Knight(final Type type, final Side side) {
-        super(type, side);
-        this.movePatterns = Arrays.asList(KnightMovePattern.values());
+    public Knight(final Side side) {
+        this.immediatePiece = new ImmediatePiece(Type.KNIGHT, side, List.of(KnightMovePattern.values()));
     }
 
     @Override
-    protected void validate(final Type type, final Side side) {
-        validateType(type);
-        validateSide(side);
-    }
-
-    private void validateType(final Type type) {
-        if (type != Type.KNIGHT) {
-            throw new IllegalArgumentException("나이트의 타입이 잘못되었습니다.");
-        }
-    }
-
-    private void validateSide(final Side side) {
-        if (side == Side.NEUTRALITY) {
-            throw new IllegalArgumentException("나이트는 중립적인 기물이 아닙니다.");
-        }
+    public List<Position> findMovablePosition(final Position source, final Board board) {
+        return immediatePiece.findMovablePosition(source, board);
     }
 
     @Override
-    protected List<MovePattern> getMovePatterns() {
-        return movePatterns;
+    public String name() {
+        return immediatePiece.name();
+    }
+
+    @Override
+    public Side side() {
+        return immediatePiece.side();
+    }
+
+    @Override
+    public boolean isPawn() {
+        return immediatePiece.isPawn();
+    }
+
+    @Override
+    public void changePawnMoveState() {
+        immediatePiece.changePawnMoveState();
     }
 }
