+++ src/main/java/chess/domain/piece/LinearPiece.java
@@ -6,32 +6,85 @@ import chess.domain.position.Position;
 import java.util.ArrayList;
 import java.util.List;
 
-public abstract class LinearPiece extends Piece {
+public final class LinearPiece implements Piece {
 
-    public LinearPiece(final Type type, final Side side) {
-        super(type, side);
+    private final Type type;
+    private final Side side;
+    private final List<MovePattern> movePatterns;
+
+    public LinearPiece(final Type type, final Side side, final List<MovePattern> movePatterns) {
+        this.type = type;
+        this.side = side;
+        this.movePatterns = movePatterns;
     }
 
     @Override
-    public List<Position> findMovablePositions(final Position source, final Board board) {
+    public List<Position> findMovablePosition(final Position source, final Board board) {
         final List<Position> movablePositions = new ArrayList<>();
-        final List<MovePattern> movePatterns = getMovePatterns();
-        for (MovePattern movePattern : movePatterns) {
-            Position nextPosition = source;
-            boolean killFlag = false;
-            while (isRangeValid(nextPosition, movePattern) && !killFlag) {
-                nextPosition = nextPosition.move(movePattern);
-                killFlag = checkSide(movablePositions, nextPosition, board);
-            }
+
+        for (final MovePattern movePattern : movePatterns) {
+            movablePositions.addAll(
+                    findMovablePositionByMovePattern(source, movePattern, board)
+            );
         }
+
         return movablePositions;
     }
 
-    private boolean checkSide(final List<Position> movablePositions, final Position nextPosition, final Board board) {
-        final Side nextSide = board.findSideByPosition(nextPosition);
-        if (nextSide != this.side) {
-            movablePositions.add(nextPosition);
+    private List<Position> findMovablePositionByMovePattern(
+            final Position source,
+            final MovePattern movePattern,
+            final Board board) {
+
+        List<Position> movablePosition = new ArrayList<>();
+        Position currentPosition = source;
+        while (canMoveMore(currentPosition, movePattern, board)) {
+            Position nextPosition = currentPosition.move(movePattern);
+            movablePosition.add(nextPosition);
+            currentPosition = nextPosition;
         }
-        return nextSide != this.side && nextSide != Side.NEUTRALITY;
+        return movablePosition;
+    }
+
+    private boolean canMoveMore(final Position currentPosition, final MovePattern movePattern, final Board board) {
+        final Position nextPosition = currentPosition.move(movePattern);
+        final Side currentSide = board.findSideByPosition(currentPosition);
+        final Side nextSide = board.findSideByPosition(nextPosition);
+
+        return isInRange(currentPosition, nextPosition) &&
+                isDifferentSide(currentSide, nextSide) &&
+                !hasCaptured(currentSide, nextSide);
+    }
+
+    private boolean isInRange(final Position currentPosition, final Position nextPosition) {
+        return currentPosition != nextPosition;
+    }
+
+    private boolean isDifferentSide(final Side currentSide, final Side nextSide) {
+        return currentSide != nextSide;
+    }
+
+    private boolean hasCaptured(final Side currentSide, final Side nextSide) {
+        return isDifferentSide(currentSide, nextSide) && nextSide != Side.NEUTRALITY;
+    }
+
+    @Override
+    public String name() {
+        return type.getSymbol(side);
+    }
+
+    @Override
+    public Side side() {
+        return side;
+    }
+
+    @Override
+    public boolean isPawn() {
+        return false;
+    }
+
+    @Override
+    public void changePawnMoveState() {
+        throw new UnsupportedOperationException("지원하지 않는 기능입니다.");
     }
 }
