--- src/main/java/chess/domain/piecerefactor/Bishop.java
@@ -1,37 +1,0 @@
-package chess.domain.piecerefactor;
-
-import chess.domain.board.Board;
-import chess.domain.movepattern.BishopMovePattern;
-import chess.domain.piece.Side;
-import chess.domain.piece.Type;
-import chess.domain.position.Position;
-import java.util.List;
-
-public class Bishop implements Piece {
-
-    private final LinearPiece linearPiece;
-
-    public Bishop(final Side side) {
-        this.linearPiece = new LinearPiece(Type.BISHOP, side, List.of(BishopMovePattern.values()));
-    }
-
-    @Override
-    public List<Position> findMovablePosition(final Position source, final Board board) {
-        return linearPiece.findMovablePosition(source, board);
-    }
-
-    @Override
-    public String name() {
-        return linearPiece.name();
-    }
-
-    @Override
-    public Side side() {
-        return linearPiece.side();
-    }
-
-    @Override
-    public boolean isPawn() {
-        return linearPiece.isPawn();
-    }
-}
