--- src/main/java/chess/domain/piecerefactor/Piece.java
@@ -1,17 +1,0 @@
-package chess.domain.piecerefactor;
-
-import chess.domain.board.Board;
-import chess.domain.piece.Side;
-import chess.domain.position.Position;
-import java.util.List;
-
-public interface Piece {
-
-    List<Position> findMovablePosition(final Position source, final Board board);
-
-    String name();
-
-    Side side();
-
-    boolean isPawn();
-}
