--- src/test/java/chess/domain/RookTest.java
@@ -1,29 +1,0 @@
-package chess.domain;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import chess.domain.piece.Rook;
-import chess.domain.piece.Side;
-import chess.domain.piece.Type;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Test;
-
-@SuppressWarnings("NonAsciiCharacters")
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public class RookTest {
-
-    @Test
-    void 타입이_잘못되면_예외를_던진다() {
-        assertThatThrownBy(() -> new Rook(Type.EMPTY, Side.WHITE))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("룩의 타입이 잘못되었습니다.");
-    }
-
-    @Test
-    void 진영이_잘못되면_예외를_던진다() {
-        assertThatThrownBy(() -> new Rook(Type.ROOK, Side.NEUTRALITY))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("룩은 중립적인 기물이 아닙니다.");
-    }
-}
